fix: agressively pin transitive deps in Cargo.toml
Since we patch ring over a version of 0.16.10,
any transitive dep we have which wants a higher version of ring
will break out build. Cargo is happy to include two different
versions of ring in our build. To allow cargo to build the lock
file (instead of having to do manual surgery on it), we must
agressively pin the several transitive deps to prevent cargo
from selecting a version of ring other than 0.16.10.